requirements.txt
gunicorn
flask
newspaper3k
nltk
textblob
validators
requests